Reducing targets for test_create_pool_quantity to support 200 pols.
Features: pool_create_tests
