+++ src/tests/ftest/pool/create_capacity_test.yaml
@@ -10,14 +10,12 @@ hosts:
   test_clients:
     - client-H
 timeouts:
-  test_create_max_pool_scm_only: 180
-  test_create_max_pool: 300
-  test_create_no_space: 300
-  test_create_no_space_loop: 1200
+  test_create_pool_quantity: 600
 server_config:
   name: daos_server
   servers:
     0:
+      targets: 4
       bdev_class: nvme
       bdev_list: ["aaaa:aa:aa.a", "bbbb:bb:bb.b"]
       scm_class: dcpm
