+++ src/tests/ftest/pool/create_test.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 """
-(C) Copyright 2020 Intel Corporation.
+(C) Copyright 2021 Intel Corporation.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -21,10 +21,7 @@ provided in Contract No. B609815.
 Any reproduction of computer software, computer software documentation, or
 portions thereof marked with this legend must also reproduce the markings.
 """
-import time
-
 from pool_test_base import PoolTestBase
-from server_utils import ServerFailed
 
 
 class PoolCreateTests(PoolTestBase):
@@ -36,31 +33,6 @@ class PoolCreateTests(PoolTestBase):
 
     :avocado: recursive
     """
-
-    def check_pool_creation(self, max_duration):
-        """Check the duration of each pool creation meets the requirement.
-
-        Args:
-            max_duration (int): max pool creation duration allowed in seconds
-
-        """
-        durations = []
-        for index, pool in enumerate(self.pool):
-            start = float(time.time())
-            pool.create()
-            durations.append(float(time.time()) - start)
-            self.log.info(
-                "Pool %s creation: %s seconds", index + 1, durations[-1])
-
-        exceeding_duration = 0
-        for index, duration in enumerate(durations):
-            if duration > max_duration:
-                exceeding_duration += 1
-
-        self.assertEqual(
-            exceeding_duration, 0,
-            "Pool creation took longer than {} seconds on {} pool(s)".format(
-                max_duration, exceeding_duration))
 
     def test_create_max_pool_scm_only(self):
         """JIRA ID: DAOS-5114 / SRS-1.
@@ -95,60 +67,6 @@ class PoolCreateTests(PoolTestBase):
         # Create 1 pool using 90% of the available capacity
         self.pool = self.get_pool_list(1, 0.9, 0.9, 1)
         self.check_pool_creation(120)
-
-    def test_create_pool_quantity(self):
-        """JIRA ID: DAOS-5114 / SRS-2 / SRS-4.
-
-        Test Description:
-            Create 200 pools on all of the servers.
-            Perform an orderly system shutdown via cmd line (dmg).
-            Restart the system via cmd line tool (dmg).
-            Verify that DAOS is ready to accept requests with in 2 minutes.
-
-        :avocado: tags=all,pr,daily_regression
-        :avocado: tags=hw,large
-        :avocado: tags=pool
-        :avocado: tags=pool_create_tests,create_performance
-        """
-        # Create some number of pools each using a equal amount of 60% of the
-        # available capacity, e.g. 0.6% for 100 pools.
-        quantity = self.params.get("quantity", "/run/pool/*", 1)
-        ratio = 0.6 / quantity
-        self.pool = self.get_pool_list(quantity, ratio, ratio, 1)
-        self.check_pool_creation(3)
-
-        # Verify DAOS can be restarted in less than 2 minutes
-        try:
-            self.server_managers[0].system_stop()
-        except ServerFailed as error:
-            self.fail(error)
-
-        start = float(time.time())
-        try:
-            self.server_managers[0].system_start()
-        except ServerFailed as error:
-            self.fail(error)
-
-        duration = float(time.time()) - start
-        self.assertLessEqual(
-            duration, 120,
-            "DAOS not ready to accept requests with in 2 minutes")
-
-        # Verify all the pools exists after the restart
-        detected_pools = [uuid.lower() for uuid in self.dmg.pool_list()]
-        missing_pools = []
-        for pool in self.pool:
-            pool_uuid = pool.uuid.lower()
-            if pool_uuid not in detected_pools:
-                missing_pools.append(pool_uuid)
-        if missing_pools:
-            self.fail(
-                "The following created pools were not detected in the pool "
-                "list after rebooting the servers:\n  [{}]: {}".format(
-                    len(missing_pools), ", ".join(missing_pools)))
-        self.assertEqual(
-            len(self.pool), len(detected_pools),
-            "Additional pools detected after rebooting the servers")
 
     def test_create_no_space(self):
         """JIRA ID: DAOS-3728.
